Lookup abstractions refactor
