+++ nifi-ignite-extensions/nifi-ignite-services/src/main/resources/META-INF/services/org.apache.nifi.controller.ControllerService
@@ -1,3 +1,3 @@
-pl.touk.nifi.services.IgniteLookupService
+pl.touk.nifi.services.IgniteRecordLookupService
 pl.touk.nifi.services.IgniteDBCPConnectionPool
 pl.touk.nifi.services.IgniteDistributedMapCacheClient
